+++ service-implementation/src/main/java/com/liaison/mailbox/rtdm/dao/RuntimeProcessorsDAOBase.java
@@ -14,13 +14,16 @@ import com.liaison.commons.jpa.DAOUtil;
 import com.liaison.commons.jpa.GenericDAOBase;
 import com.liaison.commons.util.UUIDGen;
 import com.liaison.mailbox.MailBoxConstants;
+import com.liaison.mailbox.enums.Messages;
 import com.liaison.mailbox.rtdm.model.ProcessorExecutionState;
 import com.liaison.mailbox.rtdm.model.RuntimeProcessors;
 import com.liaison.mailbox.service.core.fsm.ProcessorExecutionStateDTO;
+import com.liaison.mailbox.service.exception.MailBoxConfigurationServicesException;
 import com.liaison.mailbox.service.util.MailBoxUtil;
 
 import javax.persistence.EntityManager;
 import javax.persistence.EntityTransaction;
+import javax.ws.rs.core.Response;
 
 import java.util.Date;
 import java.util.List;
@@ -95,7 +98,7 @@ public class RuntimeProcessorsDAOBase extends GenericDAOBase<RuntimeProcessors> 
     }
 
     @Override
-    public void addProcessor(ProcessorExecutionStateDTO executionStateDTO) {
+    public void addProcessor(ProcessorExecutionStateDTO executionStateDTO, String clusterType) {
 
         RuntimeProcessors processors = new RuntimeProcessors();
         processors.setPguid(UUIDGen.getCustomUUID());
@@ -111,9 +114,51 @@ public class RuntimeProcessorsDAOBase extends GenericDAOBase<RuntimeProcessors> 
         prcsrExecution.setProcessors(processors);
         prcsrExecution.setOriginatingDc(DATACENTER_NAME);
         processors.setProcessorExecState(prcsrExecution);
-        processors.setClusterType(MailBoxUtil.CLUSTER_TYPE);
+        processors.setClusterType(clusterType);
 
         persist(processors);
+    }
+    
+    /**
+     * Update the Runtimeprocessors clusterType by processorId
+     * 
+     * @param clusterType
+     * @param processorId
+     */
+    @Override
+    public void updateClusterType(String clusterType, String processorId) {
+        
+        EntityManager entityManager = null;
+        EntityTransaction tx = null;
+        try {
+            
+            entityManager = DAOUtil.getEntityManager(persistenceUnitName);
+            tx = entityManager.getTransaction();
+            tx.begin();
+            
+            //update the Processors ClusterType
+            int count = entityManager.createNativeQuery(UPDATES_PROCESSOR_CLUSTERTYPE)
+                        .setParameter(MailBoxConstants.CLUSTER_TYPE, clusterType)
+                        .setParameter(PROCESSOR_ID, processorId)
+                        .executeUpdate();
+            
+            if (count == 0) {
+                throw new MailBoxConfigurationServicesException(Messages.INVALID_PROCESSOR_ID, processorId,
+                        Response.Status.BAD_REQUEST);
+            }
+            //commits the transaction
+            tx.commit();
+        
+        } catch (Exception e) {
+            if (null != tx && tx.isActive()) {
+                tx.rollback();
+            }
+            throw e;
+        } finally {
+            if (null != entityManager) {
+                entityManager.close();
+            }
+        }
     }
 
 }